Abstract most of extract to make adding new marketplaces easier
